feat(patch): applied xtile
Xtile is like flextile, in the sense that it's not a particular layout,
but a way of customizing the layout strategies by changing the relation
between the main and stack areas and the relation between clients
either in the main or stack areas.

Unlike flextile it also comes with a kind of cfacts of its own. But it
only affects the first (or last?) client in each area, as opposed to just
any client you want.

Unlike flextile it doesn't have a deck style layout. The author seems to
have a rather strong opinion, and believes micromanaging the order of
clients in the stack is an anti-pattern. The deck layout has an annoying
quirk of leaving a different client on top, depending on whether you
leave the deck using the j or k key. The author's suggestion is to
rather use tags if you want to hide a client, which is excellent advice.

Not only can you select what tags to assign a client to, you can also
select whatever combination of tags you want in view. So you can assign
just one, or maybe a few, clients to each tag, and then select and
deselect whatever combination of tags, depending on what clients you
want in view.