style: change new lines for some braces
Some patch authors put control structure braces on new lines, which is
ugly. Others put the opening brace of functions on the same line, which
is fine but not consistent with the original code. Besides, vim's native
function movements (`[[`,`]]` etc) seem the only work when the brace is
on a new line, so that's what I'm going with.