!feat: launch script renames and improvements
Includes new hacks to keep the terminal around after the original CLI
program is closed. Useful when I launch the `nnn` terminal file manager
and press crtl+g to cd-on-quit into the opened directory. Previously, if
`nnn` was launched through a DWM shortcut, the terminal would then just
close because the command was finished. Now it spawns a regular bash
session instead. So now I can use cd-on-quit like I'm used to when
launching `nnn` through a hotkey.

Also adapts to renamed scripts. I changed `launch_cmus.sh` (and others)
to `st_cmus.sh`, to make it more clear that they spawn a terminal.

In a way, this pattern launches interactive CLI programs from a
non-interactive context. Like when called by a window manager or dmenu.